#66: Higher MCD is better, just like every other stat.

+++ aigilas/src/sps/core/Settings.java
@@ -32,8 +32,8 @@ public class Settings {
     public final int maxRoomCount;
     public final int wallDecayPercent;
     public final int enemyStatMultiplier;
-    public final float defaultSpeed;
-    public final float defaultRegen;
+    public final int defaultSpeed;
+    public final int defaultRegen;
     public final float turnTime;
 
     public final boolean musicEnabled;
@@ -95,8 +95,8 @@ public class Settings {
         maxRoomCount = getInt(_settings.get("maxRoomCount"));
         wallDecayPercent = getInt(_settings.get("wallDecayPercent"));
         enemyStatMultiplier = getInt(_settings.get("enemyStatMultiplier"));
-        defaultSpeed = getFloat(_settings.get("defaultSpeed"));
-        defaultRegen = getFloat(_settings.get("defaultRegen"));
+        defaultSpeed = getInt(_settings.get("defaultSpeed"));
+        defaultRegen = getInt(_settings.get("defaultRegen"));
         turnTime = 1 / getFloat(_settings.get("turnsPerSecond"));
 
         // Audio
